Closes #2:  Automatically pull in data-private updates
This changeset adds support for pulling in updates from the data-private repo so that when we update the Team API site (team-api.18f.gov) we are retrieving the latest information.  This is important in the cases of when we have projects who do not have their own GitHub repo - the data-private repo is the only source to retrieve information about them.